refactor: api_calls 함수에서 중복되는 요소들을 colect_data.py로 이동
